backend/new_1/requirements.txt
fastapi
uvicorn
python-dotenv
pillow
transformers
torch
numpy
pinecone
python-jose[cryptography]
python-multipart
jose
